// src/types/rickAndMorty.ts
// src/types/rickAndMorty.ts

export interface CharacterOrigin {
    name: string;
    url: string;
  }
  
  export interface CharacterLocation {
    name: string;
    url: string;
  }
  
  export interface Character {
    id: number;
    name: string;
    status: "Alive" | "Dead" | "unknown";
    species: string;
    type: string;
    gender: "Female" | "Male" | "Genderless" | "unknown";
    origin: CharacterOrigin;
    location: CharacterLocation;
    image: string;
    episode: string[];
    url: string;
    created: string;
  }
  
  export interface CharacterListResponse {
    info: {
      count: number;
      pages: number;
      next: string | null;
      prev: string | null;
    };
    results: Character[];
  }
  